requirements.txt
fastapi
uvicorn
spacy
python-dotenv
openai
greenlet>=2.0
sqlalchemy
asyncpg
python-jose
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl